+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/heartbeat_method/mod.rs
@@ -2,12 +2,13 @@ use cdk_framework::act::node::canister_method::{CanisterMethodType, HeartbeatMet
 
 mod rust;
 
-use crate::{canister_method::errors, ts_ast::AzleFnDecl, TsAst};
+use crate::{canister_method::errors, ts_ast::azle_program::HelperMethods, TsAst};
 
 impl TsAst {
     pub fn build_heartbeat_method(&self) -> Option<HeartbeatMethod> {
-        let heartbeat_fn_decls: Vec<AzleFnDecl> =
-            self.get_annotated_function_decls_of_type(CanisterMethodType::Heartbeat);
+        let heartbeat_fn_decls = self
+            .azle_programs
+            .get_azle_fn_decls_of_type(CanisterMethodType::Heartbeat);
 
         if heartbeat_fn_decls.len() > 1 {
             let error_message = errors::build_duplicate_method_types_error_message(
